revert clang comment fix

+++ src/fheroes2/agg/agg_image.cpp
@@ -1503,7 +1503,7 @@ namespace fheroes2
                     // The evil buttons' released state are 2 pixels wider.
                     const int offsetEvilX = isEvilInterface ? 2 : 0;
 
-                    for ( size_t i = 0; i < 4; ++i ) {
+                    for ( int32_t i = 0; i < 4; ++i ) {
                         // released
                         const Sprite & originalReleased = GetICN( originalIcnId, 2 * i );
 
@@ -1540,7 +1540,7 @@ namespace fheroes2
                 const int baseIcnId = ICN::X_CMPBTN;
 
                 if ( useOriginalResources() ) {
-                    for ( size_t i = 0; i < 4; ++i ) {
+                    for ( int32_t i = 0; i < 4; ++i ) {
                         // released
                         const Sprite & originalReleased = GetICN( baseIcnId, 2 * i );
